+++ test/features/exercises/data/datasources/exercises_datasource_test.dart
@@ -3,14 +3,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:pop_network/pop_network.dart';
 
-// 1. Definição do Mock
 class IApiManagerMock extends Mock implements IApiManager {}
 
 void main() {
   late IApiManager apiManager;
   late ExercisesDatasource dataSource;
 
-  // 2. Resposta de sucesso simulada
   final tResponse = Response(
     data: [
       {'name': 'Push up'},
@@ -29,12 +27,11 @@ void main() {
       'Deve retornar um Response quando a chamada ao API Manager for bem sucedida',
       () async {
         // ARRANGE
-        // O segredo para não dar erro de Null:
-        // Use any() para a String do path e any(named: ...) para o Map de queryParameters.
         when(
           () => apiManager.get(
             any(),
             queryParameters: any(named: 'queryParameters'),
+            mockReplyParams: any(named: 'mockReplyParams'),
           ),
         ).thenAnswer((_) async => tResponse);
 
@@ -44,12 +41,11 @@ void main() {
         // ASSERT
         expect(result, isA<Response>());
         expect(result.statusCode, 200);
-
-        // Verifica se o datasource enviou o path correto para a PopNetwork
         verify(
           () => apiManager.get(
             '/v1/exercises',
             queryParameters: any(named: 'queryParameters'),
+            mockReplyParams: any(named: 'mockReplyParams'),
           ),
         ).called(1);
       },
@@ -67,6 +63,7 @@ void main() {
         () => apiManager.get(
           any(),
           queryParameters: any(named: 'queryParameters'),
+          mockReplyParams: any(named: 'mockReplyParams'),
         ),
       ).thenAnswer((_) async => tErrorResponse);